backend/receipt_ai/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
Pillow==11.2.1
torch==2.8.0
transformers==4.55.0
accelerate==1.10.1
